requirements.txt
openai
anthropic
google-genai
colorama
tiktoken
markdown